requirements.txt
flask
faiss-cpu
requests